+++ src/zandronum/src/g_hexen/a_iceguy.cpp
@@ -37,8 +37,8 @@ DEFINE_ACTION_FUNCTION(AActor, A_IceGuyLook)
 		an = (self->angle+ANG90)>>ANGLETOFINESHIFT;
 
 		mo = Spawn (WispTypes[pr_iceguylook()&1],
-			self->x+FixedMul(dist, finecosine[an]),
-			self->y+FixedMul(dist, finesine[an]),
+			self->x+FixedMul(dist, finecosine[(int)(an)]),
+			self->y+FixedMul(dist, finesine[(int)(an)]),
 			self->z+60*FRACUNIT, ALLOW_REPLACE);
 
 		// [RK] Clients spawn these on their own. In order to prevent the 
@@ -71,8 +71,8 @@ DEFINE_ACTION_FUNCTION(AActor, A_IceGuyChase)
 		an = (self->angle+ANG90)>>ANGLETOFINESHIFT;
 
 		mo = Spawn (WispTypes[pr_iceguychase()&1],
-			self->x+FixedMul(dist, finecosine[an]),
-			self->y+FixedMul(dist, finesine[an]),
+			self->x+FixedMul(dist, finecosine[(int)(an)]),
+			self->y+FixedMul(dist, finesine[(int)(an)]),
 			self->z+60*FRACUNIT, ALLOW_REPLACE);
 		if (mo)
 		{
@@ -115,13 +115,13 @@ DEFINE_ACTION_FUNCTION(AActor, A_IceGuyAttack)
 	}
 	an = (self->angle+ANG90)>>ANGLETOFINESHIFT;
 	P_SpawnMissileXYZ(self->x+FixedMul(self->radius>>1,
-		finecosine[an]), self->y+FixedMul(self->radius>>1,
-		finesine[an]), self->z+40*FRACUNIT, self, self->target,
+		finecosine[(int)(an)]), self->y+FixedMul(self->radius>>1,
+		finesine[(int)(an)]), self->z+40*FRACUNIT, self, self->target,
 		PClass::FindClass ("IceGuyFX"), true, NULL, true);	// [BB] Inform the clients.
 	an = (self->angle-ANG90)>>ANGLETOFINESHIFT;
 	P_SpawnMissileXYZ(self->x+FixedMul(self->radius>>1,
-		finecosine[an]), self->y+FixedMul(self->radius>>1,
-		finesine[an]), self->z+40*FRACUNIT, self, self->target,
+		finecosine[(int)(an)]), self->y+FixedMul(self->radius>>1,
+		finesine[(int)(an)]), self->z+40*FRACUNIT, self, self->target,
 		PClass::FindClass ("IceGuyFX"), true, NULL, true);	// [BB] Inform the clients.
 	S_Sound (self, CHAN_WEAPON, self->AttackSound, 1, ATTN_NORM, true);	// [BB] Inform the clients.
 }
