+++ src/zandronum/src/r_data/r_translate.cpp
@@ -330,7 +330,7 @@ void FRemapTable::AddIndexRange(int start, int end, int pal1, int pal2)
 	palstep = ((pal2 << FRACBITS) - palcol) / (end - start);
 	for (int i = start; i <= end; palcol += palstep, ++i)
 	{
-		int j = GPalette.Remap[i], k = GPalette.Remap[palcol >> FRACBITS];
+		int j = GPalette.Remap[i], k = GPalette.Remap[(int)(palcol >> FRACBITS)];
 		Remap[j] = k;
 		Palette[j] = GPalette.BaseColors[k];
 		Palette[j].a = j == 0 ? 0 : 255;
